Update : Définition méthode achats (presque finie, voir TODO) Fix : prixTTC dans vente Fix : souscrire fidelité, chaque client fidele sera ajouté a un dictionnaire dans le main (plus pratique pour accéder aux attributs).

+++ classe/vente/vente.hpp
@@ -10,21 +10,21 @@ La vente avec réduction se fait avec la méthode virtuelle
 class Vente {
 private:
     Produit m_produit;           // Produit associé à la vente
-    int m_prixTTC = 0;           // Prix total TTC
+    double m_prixTTC = 0;        // Prix total TTC
     int m_quantite = 0;          // Quantité vendue
     std::string m_dateVente;     // Date de la vente (format jj/mm/aaaa)
 
 public:
     // Constructeur
-    Vente(const Produit& produit, int prix, int quantite, const std::string& dateVente);
+    Vente(const Produit& produit, double prix, int quantite, const std::string& dateVente);
 
     // Setters
-    void setPrixTTC(int prixTTC);
+    void setPrixTTC(double prixTTC);
     void setQuantite(int quantite);
     void setDateVente(const std::string& dateVente);
 
     // Getters
-    int getPrixTTC() const;
+    double getPrixTTC() const;
     int getQuantite() const;
     const std::string& getDateVente() const;
